[UK] Script for 2021 England council changes.
No need to worry about which council wards go to which unitary council,
as find_parents will sort that for us.
